ml-server/requirements.txt
flask
transformers
torch
Pillow
sentence-transformers
Flask-Cors
